TGP-2890: move nav files for categorisation
