+++ test/models/CategorisationAnswersSpec.scala
@@ -20,7 +20,6 @@ import base.SpecBase
 import base.TestConstants.testRecordId
 import models.ott.CategorisationInfo
 import org.scalatest.Inside.inside
-import pages._
 import pages.categorisation.{AssessmentPage, HasSupplementaryUnitPage, ReassessmentPage, SupplementaryUnitPage}
 import queries.{CategorisationDetailsQuery, LongerCategorisationDetailsQuery}
